Reorganized and slightly reformatted
