+++ OLMoE.swift/Model/Context.swift
@@ -1,0 +1,23 @@
+//
+//  Context.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+
+
+import Foundation
+import llama
+
+public class Context {
+    let pointer: OpaquePointer
+    init(_ model: Model, _ params: llama_context_params) {
+        self.pointer = llama_new_context_with_model(model, params)
+    }
+    deinit {
+        llama_free(pointer)
+    }
+    func decode(_ batch: llama_batch) {
+        guard llama_decode(pointer, batch) == 0 else { fatalError("llama_decode failed") }
+    }
+}
