+++ OLMoE.swift/Model/LLM.swift
@@ -445,373 +445,3 @@ open class LLM: ObservableObject {
         model.encode(text)
     }
 }
-
-extension Model {
-    public var endToken: Token { llama_token_eos(self) }
-    public var newLineToken: Token { llama_token_nl(self) }
-    
-    public func shouldAddBOS() -> Bool {
-        let addBOS = llama_add_bos_token(self);
-        guard !addBOS else {
-            return llama_vocab_type(self) == LLAMA_VOCAB_TYPE_SPM
-        }
-        return addBOS
-    }
-    
-    public func decodeOnly(_ token: Token) -> String {
-        var nothing: [CUnsignedChar] = []
-        return decode(token, with: &nothing)
-    }
-    
-    public func decode(_ token: Token, with multibyteCharacter: inout [CUnsignedChar]) -> String {
-        var bufferLength = 16
-        var buffer: [CChar] = .init(repeating: 0, count: bufferLength)
-        let actualLength = Int(llama_token_to_piece(self, token, &buffer, Int32(bufferLength), 0, false))
-        guard 0 != actualLength else { return "" }
-        if actualLength < 0 {
-            bufferLength = -actualLength
-            buffer = .init(repeating: 0, count: bufferLength)
-            llama_token_to_piece(self, token, &buffer, Int32(bufferLength), 0, false)
-        } else {
-            buffer.removeLast(bufferLength - actualLength)
-        }
-        if multibyteCharacter.isEmpty, let decoded = String(cString: buffer + [0], encoding: .utf8) {
-            return decoded
-        }
-        multibyteCharacter.append(contentsOf: buffer.map { CUnsignedChar(bitPattern: $0) })
-        guard let decoded = String(data: .init(multibyteCharacter), encoding: .utf8) else { return "" }
-        multibyteCharacter.removeAll(keepingCapacity: true)
-        return decoded
-    }
-
-    public func encode(_ text: borrowing String) -> [Token] {
-        let addBOS = true
-        let count = Int32(text.cString(using: .utf8)!.count)
-        var tokenCount = count + 1
-        let cTokens = UnsafeMutablePointer<llama_token>.allocate(capacity: Int(tokenCount)); defer { cTokens.deallocate() }
-        tokenCount = llama_tokenize(self, text, count, cTokens, tokenCount, addBOS, false)
-        let tokens = (0..<Int(tokenCount)).map { cTokens[$0] }
-        
-        print("Encoded tokens: \(tokens)")  // Add this line to log the resulting tokens
-        
-        return tokens
-    }
-}
-
-private class Context {
-    let pointer: OpaquePointer
-    init(_ model: Model, _ params: llama_context_params) {
-        self.pointer = llama_new_context_with_model(model, params)
-    }
-    deinit {
-        llama_free(pointer)
-    }
-    func decode(_ batch: llama_batch) {
-        guard llama_decode(pointer, batch) == 0 else { fatalError("llama_decode failed") }
-    }
-}
-
-extension llama_batch {
-    mutating func clear() {
-        self.n_tokens = 0
-    }
-    
-    mutating func add(_ token: Token, _ position: Int32, _ ids: [Int], _ logit: Bool) {
-        let i = Int(self.n_tokens)
-        self.token[i] = token
-        self.pos[i] = position
-        self.n_seq_id[i] = Int32(ids.count)
-        if let seq_id = self.seq_id[i] {
-            for (j, id) in ids.enumerated() {
-                seq_id[j] = Int32(id)
-            }
-        }
-        self.logits[i] = logit ? 1 : 0
-        self.n_tokens += 1
-    }
-}
-
-extension [String] {
-    mutating func scoup(_ count: Int) {
-        guard 0 < count else { return }
-        let firstIndex = count
-        let lastIndex = count * 2
-        self.removeSubrange(firstIndex..<lastIndex)
-    }
-}
-
-extension Token {
-    enum Kind {
-        case end
-        case couldBeEnd
-        case normal
-    }
-}
-
-public enum Role {
-    case user
-    case bot
-}
-
-public struct Template {
-    public typealias Attachment = (prefix: String, suffix: String)
-    public let system: Attachment
-    public let user: Attachment
-    public let bot: Attachment
-    public let systemPrompt: String?
-    public let stopSequence: String?
-    public let prefix: String
-    public let shouldDropLast: Bool
-    
-    public init(
-        prefix: String = "",
-        system: Attachment? = nil,
-        user: Attachment? = nil,
-        bot: Attachment? = nil,
-        stopSequence: String? = nil,
-        systemPrompt: String?,
-        shouldDropLast: Bool = false
-    ) {
-        self.system = system ?? ("", "")
-        self.user = user  ?? ("", "")
-        self.bot = bot ?? ("", "")
-        self.stopSequence = stopSequence
-        self.systemPrompt = systemPrompt
-        self.prefix = prefix
-        self.shouldDropLast = shouldDropLast
-    }
-    
-    public var preprocess: (_ input: String, _ history: [Chat]) -> String {
-        return { [self] input, history in
-            var processed = prefix
-            if let systemPrompt {
-                processed += "\(system.prefix)\(systemPrompt)\(system.suffix)"
-            }
-            for chat in history {
-                if chat.role == .user {
-                    processed += "\(user.prefix)\(chat.content)\(user.suffix)"
-                } else {
-                    processed += "\(bot.prefix)\(chat.content)\(bot.suffix)"
-                }
-            }
-            processed += "\(user.prefix)\(input)\(user.suffix)"
-            if shouldDropLast {
-                processed += bot.prefix.dropLast()
-            } else {
-                processed += bot.prefix
-            }
-            return processed
-        }
-    }
-    
-    
-    public static func OLMoE(_ systemPrompt: String? = nil) -> Template {
-        return Template(
-            prefix: "<|endoftext|>",
-            system: ("<|system|>\n", "\n"),
-            user: ("<|user|>\n", "\n"),
-            bot: ("<|assistant|>\n", "\n"),
-            stopSequence: "<|endoftext|>",
-            systemPrompt: systemPrompt
-        )
-    }
-    
-    public static func chatML(_ systemPrompt: String? = nil) -> Template {
-        return Template(
-            system: ("<|im_start|>system\n", "<|im_end|>\n"),
-            user: ("<|im_start|>user\n", "<|im_end|>\n"),
-            bot: ("<|im_start|>assistant\n", "<|im_end|>\n"),
-            stopSequence: "<|im_end|>",
-            systemPrompt: systemPrompt
-        )
-    }
-    
-    public static func alpaca(_ systemPrompt: String? = nil) -> Template {
-        return Template(
-            system: ("", "\n\n"),
-            user: ("### Instruction:\n", "\n\n"),
-            bot: ("### Response:\n", "\n\n"),
-            stopSequence: "###",
-            systemPrompt: systemPrompt
-        )
-    }
-    
-    public static func llama(_ systemPrompt: String? = nil) -> Template {
-        return Template(
-            prefix: "[INST] ",
-            system: ("<<SYS>>\n", "\n<</SYS>>\n\n"),
-            user: ("", " [/INST]"),
-            bot: (" ", "</s><s>[INST] "),
-            stopSequence: "</s>",
-            systemPrompt: systemPrompt,
-            shouldDropLast: true
-        )
-    }
-    
-    public static let mistral = Template(
-        user: ("[INST] ", " [/INST]"),
-        bot: ("", "</s> "),
-        stopSequence: "</s>",
-        systemPrompt: nil
-    )
-}
-
-public enum Quantization: String {
-    case IQ1_S
-    case IQ1_M
-    case IQ2_XXS
-    case IQ2_XS
-    case IQ2_S
-    case IQ2_M
-    case Q2_K_S
-    case Q2_K
-    case IQ3_XXS
-    case IQ3_XS
-    case IQ3_S
-    case IQ3_M
-    case Q3_K_S
-    case Q3_K_M
-    case Q3_K_L
-    case IQ4_XS
-    case IQ4_NL
-    case Q4_0
-    case Q4_1
-    case Q4_K_S
-    case Q4_K_M
-    case Q5_0
-    case Q5_1
-    case Q5_K_S
-    case Q5_K_M
-    case Q6_K
-    case Q8_0
-}
-
-public enum HuggingFaceError: Error {
-    case network(statusCode: Int)
-    case noFilteredURL
-    case urlIsNilForSomeReason
-}
-
-public struct HuggingFaceModel {
-    public let name: String
-    public let template: Template
-    public let filterRegexPattern: String
-    
-    public init(_ name: String, template: Template, filterRegexPattern: String) {
-        self.name = name
-        self.template = template
-        self.filterRegexPattern = filterRegexPattern
-    }
-    
-    public init(_ name: String, _ quantization: Quantization = .Q4_K_M, template: Template) {
-        self.name = name
-        self.template = template
-        self.filterRegexPattern = "(?i)\(quantization.rawValue)"
-    }
-    
-    func getDownloadURLStrings() async throws -> [String] {
-        let url = URL(string: "https://huggingface.co/\(name)/tree/main")!
-        let data = try await url.getData()
-        let content = String(data: data, encoding: .utf8)!
-        let downloadURLPattern = #"(?<=href=").*\.gguf\?download=true"#
-        let matches = try! downloadURLPattern.matches(in: content)
-        let root = "https://huggingface.co"
-        return matches.map { match in root + match }
-    }
-
-    func getDownloadURL() async throws -> URL? {
-        let urlStrings = try await getDownloadURLStrings()
-        for urlString in urlStrings {
-            let found = try filterRegexPattern.hasMatch(in: urlString)
-            if found { return URL(string: urlString)! }
-        }
-        return nil
-    }
-    
-    public func download(to directory: URL = .modelsDirectory, as name: String? = nil, _ updateProgress: @escaping (Double) -> Void) async throws -> URL {
-        var destination: URL
-        if let name {
-            destination = directory.appending(path: name)
-            guard !destination.exists else { updateProgress(1); return destination }
-        }
-        guard let downloadURL = try await getDownloadURL() else { throw HuggingFaceError.noFilteredURL }
-        destination = directory.appending(path: downloadURL.lastPathComponent)
-        guard !destination.exists else { return destination }
-        try await downloadURL.downloadData(to: destination, updateProgress)
-        return destination
-    }
-    
-    public static func tinyLLaMA(_ quantization: Quantization = .Q4_K_M, _ systemPrompt: String) -> HuggingFaceModel {
-        HuggingFaceModel("TheBloke/TinyLlama-1.1B-Chat-v1.0-GGUF", quantization, template: .chatML(systemPrompt))
-    }
-}
-
-extension URL {
-    @backDeployed(before: iOS 16)
-    public func appending(path: String) -> URL {
-        appendingPathComponent(path)
-    }
-    @backDeployed(before: iOS 16)
-    public static var modelsDirectory: URL {
-        let paths = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask)
-        let url = paths[0].appendingPathComponent("Models")
-        try? FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
-        
-        // Exclude from backup
-        do {
-            var mutableURL = url
-            var resourceValues = URLResourceValues()
-            resourceValues.isExcludedFromBackup = true
-            try mutableURL.setResourceValues(resourceValues)
-        } catch {
-            print("Error excluding from backup: \(error)")
-        }
-        
-        return url
-    }
-    fileprivate var exists: Bool { FileManager.default.fileExists(atPath: path) }
-    fileprivate func getData() async throws -> Data {
-        let (data, response) = try await URLSession.shared.data(from: self)
-        let statusCode = (response as! HTTPURLResponse).statusCode
-        guard statusCode / 100 == 2 else { throw HuggingFaceError.network(statusCode: statusCode) }
-        return data
-    }
-    fileprivate func downloadData(to destination: URL, _ updateProgress: @escaping (Double) -> Void) async throws {
-        var observation: NSKeyValueObservation!
-        let url: URL = try await withCheckedThrowingContinuation { continuation in
-            let task = URLSession.shared.downloadTask(with: self) { url, response, error in
-                if let error { return continuation.resume(throwing: error) }
-                guard let url else { return continuation.resume(throwing: HuggingFaceError.urlIsNilForSomeReason) }
-                let statusCode = (response as! HTTPURLResponse).statusCode
-                guard statusCode / 100 == 2 else { return continuation.resume(throwing: HuggingFaceError.network(statusCode: statusCode)) }
-                continuation.resume(returning: url)
-            }
-            observation = task.progress.observe(\.fractionCompleted) { progress, _ in
-                updateProgress(progress.fractionCompleted)
-            }
-            task.resume()
-        }
-        _ = observation
-        try FileManager.default.moveItem(at: url, to: destination)
-    }
-}
-
-extension String {
-    func matches(in content: String) throws -> [String] {
-        let pattern = try NSRegularExpression(pattern: self)
-        let range = NSRange(location: 0, length: content.utf16.count)
-        let matches = pattern.matches(in: content, range: range)
-        return matches.map { match in String(content[Range(match.range, in: content)!]) }
-    }
-    func hasMatch(in content: String) throws -> Bool {
-        let pattern = try NSRegularExpression(pattern: self)
-        let range = NSRange(location: 0, length: content.utf16.count)
-        return pattern.firstMatch(in: content, range: range) != nil
-    }
-    func firstMatch(in content: String) throws -> String? {
-        let pattern = try NSRegularExpression(pattern: self)
-        let range = NSRange(location: 0, length: content.utf16.count)
-        guard let match = pattern.firstMatch(in: content, range: range) else { return nil }
-        return String(content[Range(match.range, in: content)!])
-    }
-}
