+++ OLMoE.swift/Model/Role.swift
@@ -1,0 +1,15 @@
+//
+//  Role.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+
+
+import Foundation
+import llama
+
+public enum Role {
+    case user
+    case bot
+}
